decompose main.py - list_table_model.py
